Move files around and re-record test stubs
