+++ Books/BooksUITests/Tests/AccountUITests.swift
@@ -1,0 +1,36 @@
+//
+//  AccountUITests.swift
+//  BooksUITests
+//
+//  Created by Martin Kim Dung-Pham on 12.11.19.
+//  Copyright © 2019 Martin Kim Dung-Pham. All rights reserved.
+//
+
+import XCTest
+
+class AccountUITests: XCTestCase {
+    var app: XCUIApplication!
+
+    override func setUp() {
+        app = XCUIApplication()
+
+        let processInfo = ProcessInfo()
+        app.launchEnvironment["STUB_PATH"] = "\(processInfo.environment["PROJECT_DIR"] ?? "")/stubs"
+        app.launchEnvironment["THE_STUBBORN_NETWORK_UI_TESTING"] = "YES"
+        app.launchEnvironment["STUB_NAME"] = self.name
+
+        app.launch()
+    }
+
+    func testSignOutButton() {
+        // given
+        SignInView.signIn(app:app)
+
+        // when
+        Navigation.openAccountView(app:app)
+
+        // then
+        let signOutLabel = app.buttons["Sign out"]
+        wait(forElement:signOutLabel, timeout:5)
+    }
+}
