+++ Books/stubs/AccountUITests_testSignOutButton.json
@@ -1,0 +1,1 @@
+[{"request":{"url":"https:\/\/zones.buecherhallen.de\/app_webuser\/WebUserSvc.asmx","headerFields":["Accept-Language[:::]en-us","Accept[:::]*\/*","SOAPAction[:::]http:\/\/bibliomondo.com\/websevices\/webuser\/CheckBorrower","Content-Type[:::]text\/xml; charset=utf-8","Accept-Encoding[:::]br, gzip, deflate"],"method":"POST"},"data":"PD94bWwgdmVyc2lvbj0iMS4wIiBlbmNvZGluZz0idXRmLTgiPz48c29hcDpFbnZlbG9wZSB4bWxuczpzb2FwPSJodHRwOi8vc2NoZW1hcy54bWxzb2FwLm9yZy9zb2FwL2VudmVsb3BlLyIgeG1sbnM6eHNpPSJodHRwOi8vd3d3LnczLm9yZy8yMDAxL1hNTFNjaGVtYS1pbnN0YW5jZSIgeG1sbnM6eHNkPSJodHRwOi8vd3d3LnczLm9yZy8yMDAxL1hNTFNjaGVtYSI+PHNvYXA6Qm9keT48Q2hlY2tCb3Jyb3dlclJlc3BvbnNlIHhtbG5zPSJodHRwOi8vYmlibGlvbW9uZG8uY29tL3dlYnNldmljZXMvd2VidXNlciI+PENoZWNrQm9ycm93ZXJSZXN1bHQ+PHVzZXJJZD4xMjM0NTY3ODk8L3VzZXJJZD48cmVjb3JkPjxDaGVja0JvcnJvd2VyUmVzdWx0IHhtbG5zPSJodHRwOi8vYmlibGlvbW9uZG8uY29tL3dlYnNlcnZpY2VzL3dlYnVzZXIiPjxJZ25vcmVQaW4+MDwvSWdub3JlUGluPjxXYW50VXNlckRldGFpbHM+MDwvV2FudFVzZXJEZXRhaWxzPjxCcndyPkExMiAzNDUgNjc4IDk8L0Jyd3I+PHBpbj4qKio8L3Bpbj48QnJ3clJlY051bT44NzAwNjI8L0Jyd3JSZWNOdW0+PElzRW5jcnlwdGVkPjA8L0lzRW5jcnlwdGVkPjxDaGVja0xvZ2luVHJhcE9ubHk+MDwvQ2hlY2tMb2dpblRyYXBPbmx5PjxJbkxvZ2luRmFpbFRyYXA+MDwvSW5Mb2dpbkZhaWxUcmFwPjxUYXBObz4wPC9UYXBObz48TFJCPk9rPC9MUkI+PFZhbGlkYXRlZD4xPC9WYWxpZGF0ZWQ+PElzSW52YWxpZEJyd3JOdW0+MDwvSXNJbnZhbGlkQnJ3ck51bT48Q2F0ZWdvcnk+MjQ8L0NhdGVnb3J5PjxDYXRlZ29yeUNvZGU+TEU8L0NhdGVnb3J5Q29kZT48VXNlcklkPjEyMzQ1PC9Vc2VySWQ+PEZTSz4xODwvRlNLPjxTdGF0dXM+QWN0aXZlPC9TdGF0dXM+PFRyYXBUeXBlPjwvVHJhcFR5cGU+PHNlc3Npb25JZCB4bWxucz0iIj5hYmMtMTIzPC9zZXNzaW9uSWQ+PC9DaGVja0JvcnJvd2VyUmVzdWx0PjwvcmVjb3JkPjwvQ2hlY2tCb3Jyb3dlclJlc3VsdD48L0NoZWNrQm9ycm93ZXJSZXNwb25zZT48L3NvYXA6Qm9keT48L3NvYXA6RW52ZWxvcGU+","response":{}},{"request":{"url":"https:\/\/zones.buecherhallen.de\/app_webuser\/WebUserSvc.asmx","headerFields":["Accept-Language[:::]en-us","SOAPAction[:::]http:\/\/bibliomondo.com\/websevices\/webuser\/GetBorrowerLoans","Accept[:::]*\/*","Content-Type[:::]text\/xml; charset=utf-8","Accept-Encoding[:::]br, gzip, deflate"],"method":"POST"},"data":"PD94bWwgdmVyc2lvbj0iMS4wIiBlbmNvZGluZz0idXRmLTgiPz48c29hcDpFbnZlbG9wZSB4bWxuczpzb2FwPSJodHRwOi8vc2NoZW1hcy54bWxzb2FwLm9yZy9zb2FwL2VudmVsb3BlLyIgeG1sbnM6eHNpPSJodHRwOi8vd3d3LnczLm9yZy8yMDAxL1hNTFNjaGVtYS1pbnN0YW5jZSIgeG1sbnM6eHNkPSJodHRwOi8vd3d3LnczLm9yZy8yMDAxL1hNTFNjaGVtYSI+PHNvYXA6Qm9keT48R2V0Qm9ycm93ZXJMb2Fuc1Jlc3BvbnNlIHhtbG5zPSJodHRwOi8vYmlibGlvbW9uZG8uY29tL3dlYnNldmljZXMvd2VidXNlciI+PEdldEJvcnJvd2VyTG9hbnNSZXN1bHQ+PHVzZXJJZD4xMjM0NTY3ODk8L3VzZXJJZD48cmVjb3JkPjxHZXRCb3Jyb3dlckxvYW5zUmVzdWx0IHhtbG5zPSJodHRwOi8vYmlibGlvbW9uZG8uY29tL3dlYnNlcnZpY2VzL3dlYnVzZXIiPjxMb2FuRGV0YWlscz48L0xvYW5EZXRhaWxzPjwvR2V0Qm9ycm93ZXJMb2Fuc1Jlc3VsdD48L3JlY29yZD48L0dldEJvcnJvd2VyTG9hbnNSZXN1bHQ+PC9HZXRCb3Jyb3dlckxvYW5zUmVzcG9uc2U+PC9zb2FwOkJvZHk+PC9zb2FwOkVudmVsb3BlPg==","response":{}},{"request":{"url":"https:\/\/zones.buecherhallen.de\/app_webuser\/WebUserSvc.asmx","headerFields":["Accept-Language[:::]en-us","Accept[:::]*\/*","SOAPAction[:::]http:\/\/bibliomondo.com\/websevices\/webuser\/CheckBorrower","Content-Type[:::]text\/xml; charset=utf-8","Accept-Encoding[:::]br, gzip, deflate"],"method":"POST"},"data":"PD94bWwgdmVyc2lvbj0iMS4wIiBlbmNvZGluZz0idXRmLTgiPz48c29hcDpFbnZlbG9wZSB4bWxuczpzb2FwPSJodHRwOi8vc2NoZW1hcy54bWxzb2FwLm9yZy9zb2FwL2VudmVsb3BlLyIgeG1sbnM6eHNpPSJodHRwOi8vd3d3LnczLm9yZy8yMDAxL1hNTFNjaGVtYS1pbnN0YW5jZSIgeG1sbnM6eHNkPSJodHRwOi8vd3d3LnczLm9yZy8yMDAxL1hNTFNjaGVtYSI+PHNvYXA6Qm9keT48Q2hlY2tCb3Jyb3dlclJlc3BvbnNlIHhtbG5zPSJodHRwOi8vYmlibGlvbW9uZG8uY29tL3dlYnNldmljZXMvd2VidXNlciI+PENoZWNrQm9ycm93ZXJSZXN1bHQ+PHVzZXJJZD4xMjM0NTY3ODk8L3VzZXJJZD48cmVjb3JkPjxDaGVja0JvcnJvd2VyUmVzdWx0IHhtbG5zPSJodHRwOi8vYmlibGlvbW9uZG8uY29tL3dlYnNlcnZpY2VzL3dlYnVzZXIiPjxJZ25vcmVQaW4+MDwvSWdub3JlUGluPjxXYW50VXNlckRldGFpbHM+MDwvV2FudFVzZXJEZXRhaWxzPjxCcndyPkExMiAzNDUgNjc4IDk8L0Jyd3I+PHBpbj4qKio8L3Bpbj48QnJ3clJlY051bT44NzAwNjI8L0Jyd3JSZWNOdW0+PElzRW5jcnlwdGVkPjA8L0lzRW5jcnlwdGVkPjxDaGVja0xvZ2luVHJhcE9ubHk+MDwvQ2hlY2tMb2dpblRyYXBPbmx5PjxJbkxvZ2luRmFpbFRyYXA+MDwvSW5Mb2dpbkZhaWxUcmFwPjxUYXBObz4wPC9UYXBObz48TFJCPk9rPC9MUkI+PFZhbGlkYXRlZD4xPC9WYWxpZGF0ZWQ+PElzSW52YWxpZEJyd3JOdW0+MDwvSXNJbnZhbGlkQnJ3ck51bT48Q2F0ZWdvcnk+MjQ8L0NhdGVnb3J5PjxDYXRlZ29yeUNvZGU+TEU8L0NhdGVnb3J5Q29kZT48VXNlcklkPjEyMzQ1PC9Vc2VySWQ+PEZTSz4xODwvRlNLPjxTdGF0dXM+QWN0aXZlPC9TdGF0dXM+PFRyYXBUeXBlPjwvVHJhcFR5cGU+PHNlc3Npb25JZCB4bWxucz0iIj5hYmMtMTIzPC9zZXNzaW9uSWQ+PC9DaGVja0JvcnJvd2VyUmVzdWx0PjwvcmVjb3JkPjwvQ2hlY2tCb3Jyb3dlclJlc3VsdD48L0NoZWNrQm9ycm93ZXJSZXNwb25zZT48L3NvYXA6Qm9keT48L3NvYXA6RW52ZWxvcGU+","response":{"statusCode":200,"headerFields":["Cache-Control[:::]private, max-age=0","X-Powered-By[:::]ASP.NET","Vary[:::]Accept-Encoding","Server[:::]Microsoft-IIS\/8.0","X-AspNet-Version[:::]2.0.50727","Content-Type[:::]text\/xml; charset=utf-8","Date[:::]Tue, 12 Nov 2019 21:37:24 GMT","Content-Encoding[:::]gzip","Content-Length[:::]655"]}},{"request":{"url":"https:\/\/zones.buecherhallen.de\/app_webuser\/WebUserSvc.asmx","headerFields":["Accept[:::]*\/*","Accept-Encoding[:::]br, gzip, deflate","Content-Type[:::]text\/xml; charset=utf-8","Accept-Language[:::]en-us","SOAPAction[:::]http:\/\/bibliomondo.com\/websevices\/webuser\/GetBorrowerLoans"],"method":"POST"},"data":"PD94bWwgdmVyc2lvbj0iMS4wIiBlbmNvZGluZz0idXRmLTgiPz48c29hcDpFbnZlbG9wZSB4bWxuczpzb2FwPSJodHRwOi8vc2NoZW1hcy54bWxzb2FwLm9yZy9zb2FwL2VudmVsb3BlLyIgeG1sbnM6eHNpPSJodHRwOi8vd3d3LnczLm9yZy8yMDAxL1hNTFNjaGVtYS1pbnN0YW5jZSIgeG1sbnM6eHNkPSJodHRwOi8vd3d3LnczLm9yZy8yMDAxL1hNTFNjaGVtYSI+PHNvYXA6Qm9keT48R2V0Qm9ycm93ZXJMb2Fuc1Jlc3BvbnNlIHhtbG5zPSJodHRwOi8vYmlibGlvbW9uZG8uY29tL3dlYnNldmljZXMvd2VidXNlciI+PEdldEJvcnJvd2VyTG9hbnNSZXN1bHQ+PHVzZXJJZD4xMjM0NTY3ODk8L3VzZXJJZD48cmVjb3JkPjxHZXRCb3Jyb3dlckxvYW5zUmVzdWx0IHhtbG5zPSJodHRwOi8vYmlibGlvbW9uZG8uY29tL3dlYnNlcnZpY2VzL3dlYnVzZXIiPjxMb2FuRGV0YWlscz48L0xvYW5EZXRhaWxzPjwvR2V0Qm9ycm93ZXJMb2Fuc1Jlc3VsdD48L3JlY29yZD48L0dldEJvcnJvd2VyTG9hbnNSZXN1bHQ+PC9HZXRCb3Jyb3dlckxvYW5zUmVzcG9uc2U+PC9zb2FwOkJvZHk+PC9zb2FwOkVudmVsb3BlPg==","response":{"statusCode":200,"headerFields":["X-AspNet-Version[:::]2.0.50727","Date[:::]Tue, 12 Nov 2019 21:37:24 GMT","Server[:::]Microsoft-IIS\/8.0","Vary[:::]Accept-Encoding","Content-Length[:::]377","Cache-Control[:::]private, max-age=0","Content-Encoding[:::]gzip","Content-Type[:::]text\/xml; charset=utf-8","X-Powered-By[:::]ASP.NET"]}}]
