--- Books/stubs/BooksUITests_testSignIn.json
@@ -1,1 +1,0 @@
-[{"request":{"url":"https:\/\/zones.buecherhallen.de\/app_webuser\/WebUserSvc.asmx","headerFields":["SOAPAction[:::]http:\/\/bibliomondo.com\/websevices\/webuser\/CheckBorrower","Accept[:::]*\/*","Content-Type[:::]text\/xml; charset=utf-8","Accept-Language[:::]en-us","Accept-Encoding[:::]br, gzip, deflate"],"method":"POST"},"data":"PD94bWwgdmVyc2lvbj0iMS4wIiBlbmNvZGluZz0idXRmLTgiPz48c29hcDpFbnZlbG9wZSB4bWxuczpzb2FwPSJodHRwOi8vc2NoZW1hcy54bWxzb2FwLm9yZy9zb2FwL2VudmVsb3BlLyIgeG1sbnM6eHNpPSJodHRwOi8vd3d3LnczLm9yZy8yMDAxL1hNTFNjaGVtYS1pbnN0YW5jZSIgeG1sbnM6eHNkPSJodHRwOi8vd3d3LnczLm9yZy8yMDAxL1hNTFNjaGVtYSI+PHNvYXA6Qm9keT48Q2hlY2tCb3Jyb3dlclJlc3BvbnNlIHhtbG5zPSJodHRwOi8vYmlibGlvbW9uZG8uY29tL3dlYnNldmljZXMvd2VidXNlciI+PENoZWNrQm9ycm93ZXJSZXN1bHQ+PHVzZXJJZD4xMjM0NTY3ODk8L3VzZXJJZD48cmVjb3JkPjxDaGVja0JvcnJvd2VyUmVzdWx0IHhtbG5zPSJodHRwOi8vYmlibGlvbW9uZG8uY29tL3dlYnNlcnZpY2VzL3dlYnVzZXIiPjxJZ25vcmVQaW4+MDwvSWdub3JlUGluPjxXYW50VXNlckRldGFpbHM+MDwvV2FudFVzZXJEZXRhaWxzPjxCcndyPkExMiAzNDUgNjc4IDk8L0Jyd3I+PHBpbj4qKio8L3Bpbj48SXNFbmNyeXB0ZWQ+MDwvSXNFbmNyeXB0ZWQ+PFRhcE5vPjA8L1RhcE5vPjxMUkI+T2s8L0xSQj48VmFsaWRhdGVkPjE8L1ZhbGlkYXRlZD48SXNJbnZhbGlkQnJ3ck51bT4wPC9Jc0ludmFsaWRCcndyTnVtPjxDYXRlZ29yeT4yNDwvQ2F0ZWdvcnk+PENhdGVnb3J5Q29kZT5MRTwvQ2F0ZWdvcnlDb2RlPjxVc2VySWQ+MTIzNDU8L1VzZXJJZD48RlNLPjE4PC9GU0s+PFN0YXR1cz5BY3RpdmU8L1N0YXR1cz48VHJhcFR5cGU+PC9UcmFwVHlwZT48c2Vzc2lvbklkIHhtbG5zPSIiPmFiYy0xMjM8L3Nlc3Npb25JZD48L0NoZWNrQm9ycm93ZXJSZXN1bHQ+PC9yZWNvcmQ+PC9DaGVja0JvcnJvd2VyUmVzdWx0PjwvQ2hlY2tCb3Jyb3dlclJlc3BvbnNlPjwvc29hcDpCb2R5Pjwvc29hcDpFbnZlbG9wZT4=","response":{"statusCode":200,"headerFields":["Cache-Control[:::]private, max-age=0","X-Powered-By[:::]ASP.NET","X-AspNet-Version[:::]2.0.50727","Vary[:::]Accept-Encoding","Content-Encoding[:::]gzip","Content-Length[:::]618","Date[:::]Wed, 30 Oct 2019 19:05:46 GMT","Content-Type[:::]text\/xml; charset=utf-8","Server[:::]Microsoft-IIS\/8.0"]}},{"request":{"url":"https:\/\/zones.buecherhallen.de\/app_webuser\/WebUserSvc.asmx","headerFields":["SOAPAction[:::]http:\/\/bibliomondo.com\/websevices\/webuser\/GetBorrowerLoans","Accept[:::]*\/*","Accept-Encoding[:::]br, gzip, deflate","Accept-Language[:::]en-us","Content-Type[:::]text\/xml; charset=utf-8"],"method":"POST"},"data":"PD94bWwgdmVyc2lvbj0iMS4wIiBlbmNvZGluZz0idXRmLTgiPz48c29hcDpFbnZlbG9wZSB4bWxuczpzb2FwPSJodHRwOi8vc2NoZW1hcy54bWxzb2FwLm9yZy9zb2FwL2VudmVsb3BlLyIgeG1sbnM6eHNpPSJodHRwOi8vd3d3LnczLm9yZy8yMDAxL1hNTFNjaGVtYS1pbnN0YW5jZSIgeG1sbnM6eHNkPSJodHRwOi8vd3d3LnczLm9yZy8yMDAxL1hNTFNjaGVtYSI+PHNvYXA6Qm9keT48R2V0Qm9ycm93ZXJMb2Fuc1Jlc3BvbnNlIHhtbG5zPSJodHRwOi8vYmlibGlvbW9uZG8uY29tL3dlYnNldmljZXMvd2VidXNlciI+PEdldEJvcnJvd2VyTG9hbnNSZXN1bHQ+PHVzZXJJZD4xMjM0NTY3ODk8L3VzZXJJZD48cmVjb3JkPjxHZXRCb3Jyb3dlckxvYW5zUmVzdWx0IHhtbG5zPSJodHRwOi8vYmlibGlvbW9uZG8uY29tL3dlYnNlcnZpY2VzL3dlYnVzZXIiPjxMb2FuRGV0YWlscz48L0xvYW5EZXRhaWxzPjwvR2V0Qm9ycm93ZXJMb2Fuc1Jlc3VsdD48L3JlY29yZD48L0dldEJvcnJvd2VyTG9hbnNSZXN1bHQ+PC9HZXRCb3Jyb3dlckxvYW5zUmVzcG9uc2U+PC9zb2FwOkJvZHk+PC9zb2FwOkVudmVsb3BlPg==","response":{"statusCode":200,"headerFields":["Content-Length[:::]377","Server[:::]Microsoft-IIS\/8.0","Cache-Control[:::]private, max-age=0","Date[:::]Wed, 30 Oct 2019 19:05:46 GMT","Vary[:::]Accept-Encoding","X-AspNet-Version[:::]2.0.50727","Content-Encoding[:::]gzip","Content-Type[:::]text\/xml; charset=utf-8","X-Powered-By[:::]ASP.NET"]}}]
